s/off_t/size_t
Reported by:	antoine

+++ libpkg/utils.c
@@ -147,7 +147,7 @@ mkdirs(const char *_path)
 	return (EPKG_OK);
 }
 int
-file_to_bufferat(int dfd, const char *path, char **buffer, off_t *sz)
+file_to_bufferat(int dfd, const char *path, char **buffer, size_t *sz)
 {
 	int fd = -1;
 	struct stat st;
